requirements.txt
Flask=1.1.2
Flask-Cors=3.0.10
Flask-JWT-Extended=3.25.0
Flask-RESTful=0.3.8
Flask-SQLAlchemy=2.4.4
Flask-Uploads=0.2.1
mysql-connector-python=8.0.22
google_auth_oauthlib=0.4.2
google-api-python-client=1.12.8
SQLAlchemy=1.3.22


# send an E-mail to  "cardiology19999@gmail.com" to gain google calendar access permission